Recipe screen Available Recipes List fully functional

+++ app/src/main/java/com/example/greenplate/model/Ingredient.java
@@ -4,16 +4,16 @@ public class Ingredient {
     private String name;
     private double quantity;
     private String units;
-    private int calories;
+    private double calories;
 
-    public Ingredient(String name, double quantity, String units, int calories) {
+    public Ingredient(String name, double quantity, String units, double calories) {
         this.name = name;
         this.quantity = quantity;
         this.units = units;
         this.calories = calories;
     }
 
-    public int getCalories() { return calories;}
+    public double getCalories() { return calories;}
     public String getName() {
         return name;
     }
